Move evm tracing integration tests
